forgot git add

+++ Install.sh
@@ -1,14 +1,19 @@
 #!/bin/bash
+
+NO_CONFIRM=false
+[[ "$1" == "--no-confirm" ]] && NO_CONFIRM=true
+
 if [[ $EUID -ne 0 ]]; then
-    read -rp "  Running as user — will install to:
+    if [[ $NO_CONFIRM == false ]]; then
+        read -rp "  Running as user — will install to:
   ~/.local/bin/rsv
   ~/.local/share/bash-completion/completions/rsv
   ~/.config/fish/completions/rsv.fish
   ~/.local/share/zsh/site-functions/_rsv
   I Recommend installing as Root for sudo rsv to work
   Confirm? [y/N]: " confirm
-    [[ $confirm =~ ^[Yy]$ ]] || { echo "Cancelled."; exit 1; }
-
+        [[ $confirm =~ ^[Yy]$ ]] || { echo "Cancelled."; exit 1; }
+    fi
     mkdir -p ~/.local/bin \
              ~/.local/share/bash-completion/completions \
              ~/.config/fish/completions \
@@ -19,14 +24,15 @@ if [[ $EUID -ne 0 ]]; then
     install -m644 rsv.zsh  ~/.local/share/zsh/site-functions/_rsv
     echo "  Installed. Make sure ~/.local/bin is in your PATH."
 else
-    read -rp "  Running as root — will install to:
+    if [[ $NO_CONFIRM == false ]]; then
+        read -rp "  Running as root — will install to:
   /usr/local/bin/rsv
   /usr/share/bash-completion/completions/rsv
   /usr/share/fish/vendor_completions.d/rsv.fish
   /usr/share/zsh/site-functions/_rsv
   Confirm? [y/N]: " confirm
-    [[ $confirm =~ ^[Yy]$ ]] || { echo "Cancelled."; exit 1; }
-
+        [[ $confirm =~ ^[Yy]$ ]] || { echo "Cancelled."; exit 1; }
+    fi
     install -Dm755 rsv      /usr/local/bin/rsv
     install -Dm644 rsv.bash /usr/share/bash-completion/completions/rsv
     install -Dm644 rsv.fish /usr/share/fish/vendor_completions.d/rsv.fish
